Refactor :  works fetching logic into module
Moved the getWorks and displayWorks functions from main.js to a new getWorks.js module. Updated main.js to use the new worksInit function for initializing and displaying works, improving code organization and modularity.
